Major changes

+++ benchmark/common.hpp
@@ -12,7 +12,8 @@
 #include <zuffix/util/common.hpp>
 #include <zuffix/util/String.hpp>
 
-#include <zuffix/hash/CRC32FollyHash.hpp>
+#include <zuffix/hash/CRC32Plus32CFollyHash.hpp>
+#include <zuffix/hash/CRC32CFollyHash.hpp>
 #include <zuffix/hash/CRC32ZlibHash.hpp>
 #include <zuffix/hash/CRC32Hash.hpp>
 #include <zuffix/hash/CyclicPolyHash.hpp>
@@ -23,6 +24,7 @@
 
 #include <zuffix/EnhancedSuffixArray.hpp>
 #include <zuffix/ExactZuffixArray.hpp>
+#include <zuffix/ProbabilisticZuffixArray.hpp>
 #include <zuffix/SimpleSuffixArray.hpp>
 
 template <typename T> using CyclicPoly4Hash = zarr::CyclicPolyHash<T, 4>;
@@ -32,7 +34,7 @@ zarr::String<uint8_t> random(size_t n, const uint8_t *charset, size_t sigma) {
 	zarr::String<uint8_t> result(n);
 	static std::random_device rd;
 	static zarr::xoroshiro128plus_engine rng(rd());
-	std::uniform_int_distribution<uint8_t> dist(0, sigma - 1);
+	std::uniform_int_distribution<uint64_t> dist(0, sigma - 1);
 	for (size_t i = 0; i < n; i++) result[i] = charset[dist(rng)];
 	return result;
 }
